update website
Wave 3.6: Small update regarding the creation of a resource. Now the user must create a resource with all the media (Thumbnail, video, image, sound)
Removed the term "tracked_image" replaced by "thumbnail".
Also, it is now impossible to have a resource that does not work (no video or no thumbnail)